Separated RTD resistance / temperature model from resistance-measuring hardware
